+++ R/plot_coefs.R
@@ -22,7 +22,7 @@ plot_coefs <- function(model_input,
 
   model_input <- model_input %>%
     dplyr::mutate(bold = ifelse(sign(ui_l)*sign(ui_u) == 1, 1, .9),
-                  Coefficient = ordered(Coefficient, levels=Coefficient))  # sigh
+                  Coefficient = ordered(Coefficient, levels=Coefficient)) # sigh
   if (!is.null(trans))
     model_input <- model_input %>%
       dplyr::mutate_at(vars(value, contains('ui')), trans)