+++ tests/testthat/test_plot_gam_2d.R
@@ -9,13 +9,13 @@ library(mgcv) # you don't need this function if you don't have this package
 # example taken from the mgcv plot.gam help file.
 library(mgcv); library(dplyr)
 set.seed(0)
-d = gamSim(2, scale=.1)$data
-d$misc = rnorm(nrow(d), mean = 50, sd = 10)
+d <- gamSim(2, scale=.1)$data
+d$misc <- rnorm(nrow(d), mean = 50, sd = 10)
 b <- gam(y ~ s(x, z), data = d)
 
 
-d2 = gamSim(4)
-d2$fac_num = as.numeric(d2$fac)
+d2 <- gamSim(4)
+d2$fac_num <-  as.numeric(d2$fac)
 by_mod1 <- gam(y ~ s(x2, by=fac), data = d2)
 by_mod2 <- gam(y ~ s(x2, by=fac_num), data = d2)
 by_mod3 <- gam(y ~ s(x2, fac, bs='fs'), data = d2)
@@ -32,7 +32,7 @@ test_that('plot_gam_2d returns a ggplot',{
 })
 
 test_that('plot_gam_2d fails if not gam object',{
-  expect_error(plot_gam(lm(y ~ x*z, d), main_var = x, second_var=z))
+  expect_error(plot_gam_2d(lm(y ~ x*z, d), main_var = x, second_var=z))
 })
 
 test_that('plot_gam_2d will switch to by',{
@@ -54,8 +54,13 @@ test_that('plot_gam_by returns a ggplot',{
   expect_s3_class(plot_gam_by(by_mod1, main_var = x2, by_var = fac), 'ggplot')
 })
 
+test_that('plot_gam_by fails if not gam object',{
+  expect_error(plot_gam_by(lm(y ~ x*z, d), main_var = x, by_var = z))
+})
+
 test_that('plot_gam_by will use numeric',{
-  expect_s3_class(plot_gam_by(by_mod2, main_var = x2, by_var = fac_num), 'ggplot')
+  expect_s3_class(plot_gam_by(by_mod2, main_var = x2, by_var = fac_num),
+                  'ggplot')
 })
 
 test_that('plot_gam_by works with factor smooth',{